+++ mc/1-8---1-12/src/main/java/io/redstudioragnarok/redcore/RedCore.java
@@ -1,5 +1,6 @@
 package io.redstudioragnarok.redcore;
 
+import dev.redstudio.redcore.annotations.RedDeprecated;
 import dev.redstudio.redcore.utils.OptiNotFine;
 import io.redstudioragnarok.redcore.ticking.RedClientTicker;
 import net.minecraftforge.common.MinecraftForge;
@@ -10,13 +11,13 @@ import net.minecraftforge.fml.relauncher.SideOnly;
  * @author Luna Mira Lage (Desoroxxx)
  * @deprecated See methods for details.
  */
-@Deprecated // Todo: Remove in 0.8
+@RedDeprecated(since = "0.5", forRemovalIn = "0.8")
 public final class RedCore {
 
     /**
      * @deprecated Use {@link dev.redstudio.redcore.ticking.RedClientTicker#startClientTicker}
      */
-    @Deprecated
+    @RedDeprecated(since = "0.5", forRemovalIn = "0.8")
     @SideOnly(Side.CLIENT)
     public static void startClientTicker() {
         MinecraftForge.EVENT_BUS.register(RedClientTicker.class);
@@ -25,7 +26,7 @@ public final class RedCore {
     /**
      * @deprecated Use {@link OptiNotFine#forceOptiFineFastRenderOff}
      */
-    @Deprecated
+    @RedDeprecated(since = "0.5", forRemovalIn = "0.8")
     @SideOnly(Side.CLIENT)
     public static void forceOptiFineFastRenderOff() {
         OptiNotFine.forceOptiFineFastRenderOff();