+++ mc/1-8---1-12/src/main/java/io/redstudioragnarok/redcore/logging/RedLogger.java
@@ -1,5 +1,6 @@
 package io.redstudioragnarok.redcore.logging;
 
+import dev.redstudio.redcore.annotations.RedDeprecated;
 import org.apache.logging.log4j.Logger;
 
 import java.net.URI;
@@ -9,19 +10,19 @@ import java.util.Random;
  * @author Luna Mira Lage (Desoroxxx)
  * @deprecated See methods for details.
  */
-@Deprecated // Todo: Remove in 0.8
+@RedDeprecated(since = "0.5", forRemovalIn = "0.8")
 public final class RedLogger extends dev.redstudio.redcore.logging.RedLogger {
 
     /**
      * @deprecated Use your own random
      */
-    @Deprecated
+    @RedDeprecated(since = "0.5", forRemovalIn = "0.8")
     public static final Random RANDOM = new Random();
 
     /**
      * @deprecated Use {@link dev.redstudio.redcore.logging.RedLogger#RedLogger(String, String, Logger, String...)}
      */
-    @Deprecated
+    @RedDeprecated(since = "0.5", forRemovalIn = "0.8")
     public RedLogger(final String modName, final String newIssueLink, final Logger logger, final String... customRecomfortMessages) {
         super(modName, newIssueLink, logger, customRecomfortMessages);
     }
@@ -29,7 +30,7 @@ public final class RedLogger extends dev.redstudio.redcore.logging.RedLogger {
     /**
      * @deprecated Use {@link dev.redstudio.redcore.logging.RedLogger#RedLogger(String, URI, Logger, String...)}
      */
-    @Deprecated
+    @RedDeprecated(since = "0.5", forRemovalIn = "0.8")
     public RedLogger(final String modName, final URI newIssueLink, final Logger logger, final String... customRecomfortMessages) {
         super(modName, newIssueLink, logger, customRecomfortMessages);
     }
@@ -38,7 +39,7 @@ public final class RedLogger extends dev.redstudio.redcore.logging.RedLogger {
      * @deprecated Use {@link dev.redstudio.redcore.logging.RedLogger#framedError}
      */
     @Override
-    @Deprecated
+    @RedDeprecated(since = "0.5", forRemovalIn = "0.8")
     public void printFramedError(final String category, final String whatHappened, final String whatNow, final String... additionalInformation) {
         super.framedError(category, whatHappened, whatNow, additionalInformation);
     }