+++ mc/1-8---1-12/src/main/java/io/redstudioragnarok/redcore/ticking/RedClientTicker.java
@@ -1,5 +1,6 @@
 package io.redstudioragnarok.redcore.ticking;
 
+import dev.redstudio.redcore.annotations.RedDeprecated;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.fml.common.eventhandler.EventPriority;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
@@ -9,7 +10,7 @@ import net.minecraftforge.fml.common.gameevent.TickEvent;
  * @author Luna Mira Lage (Desoroxxx)
  * @deprecated See methods for details.
  */
-@Deprecated // Todo: Remove in 0.8
+@RedDeprecated(since = "0.5", forRemovalIn = "0.8")
 public class RedClientTicker {
 
     private static int biTickCount, pentaTickCount, decaTickCount;
@@ -17,7 +18,7 @@ public class RedClientTicker {
     /**
      * @deprecated Use {@link dev.redstudio.redcore.ticking.RedClientTicker#onClientTickEvent}
      */
-    @Deprecated
+    @RedDeprecated(since = "0.5", forRemovalIn = "0.8")
     @SubscribeEvent(priority = EventPriority.HIGHEST)
     public static void onClientTickEvent(TickEvent.ClientTickEvent clientTickEvent) {
         if (clientTickEvent.phase == TickEvent.Phase.START) {