+++ mc/1-8---1-12/src/main/java/io/redstudioragnarok/redcore/utils/MathUtil.java
@@ -1,19 +1,20 @@
 package io.redstudioragnarok.redcore.utils;
 
+import dev.redstudio.redcore.annotations.RedDeprecated;
 import dev.redstudio.redcore.math.ClampUtil;
 
 /**
  * @author Luna Mira Lage (Desoroxxx)
  * @deprecated See methods for details.
  */
-@Deprecated // Todo: Remove in 0.8
+@RedDeprecated(since = "0.5", forRemovalIn = "0.8")
 @SuppressWarnings("unused")
 public final class MathUtil {
 
     /**
      * @deprecated Use {@link ClampUtil#clampMinFirst}
      */
-    @Deprecated
+    @RedDeprecated(since = "0.5", forRemovalIn = "0.8")
     public static float clampMinFirst(final float input, final float min, final float max) {
         return ClampUtil.clampMinFirst(input, min, max);
     }
@@ -21,7 +22,7 @@ public final class MathUtil {
     /**
      * @deprecated Use {@link ClampUtil#clampMaxFirst}
      */
-    @Deprecated
+    @RedDeprecated(since = "0.5", forRemovalIn = "0.8")
     public static float clampMaxFirst(final float input, final float min, final float max) {
         return ClampUtil.clampMaxFirst(input, min, max);
     }
@@ -29,7 +30,7 @@ public final class MathUtil {
     /**
      * @deprecated Use {@link ClampUtil#clampTest}
      */
-    @Deprecated
+    @RedDeprecated(since = "0.5", forRemovalIn = "0.8")
     public static float clampTest(final float input, final float min, final float max) {
         return ClampUtil.clampTest(input, min, max);
     }
@@ -37,7 +38,7 @@ public final class MathUtil {
     /**
      * @deprecated Use {@link ClampUtil#clampMinFirst}
      */
-    @Deprecated
+    @RedDeprecated(since = "0.5", forRemovalIn = "0.8")
     public static double clampMinFirst(final double input, final double min, final double max) {
         return ClampUtil.clampMinFirst(input, min, max);
     }
@@ -45,7 +46,7 @@ public final class MathUtil {
     /**
      * @deprecated Use {@link ClampUtil#clampMaxFirst}
      */
-    @Deprecated
+    @RedDeprecated(since = "0.5", forRemovalIn = "0.8")
     public static double clampMaxFirst(final double input, final double min, final double max) {
         return ClampUtil.clampMaxFirst(input, min, max);
     }
@@ -53,7 +54,7 @@ public final class MathUtil {
     /**
      * @deprecated Use {@link ClampUtil#clampTest}
      */
-    @Deprecated
+    @RedDeprecated(since = "0.5", forRemovalIn = "0.8")
     public static double clampTest(final double input, final double min, final double max) {
         return ClampUtil.clampTest(input, min, max);
     }
@@ -61,7 +62,7 @@ public final class MathUtil {
     /**
      * @deprecated Use {@link ClampUtil#clampMinFirst}
      */
-    @Deprecated
+    @RedDeprecated(since = "0.5", forRemovalIn = "0.8")
     public static byte clampMinFirst(final byte input, final byte min, final byte max) {
         return ClampUtil.clampMinFirst(input, min, max);
     }
@@ -69,7 +70,7 @@ public final class MathUtil {
     /**
      * @deprecated Use {@link ClampUtil#clampMaxFirst}
      */
-    @Deprecated
+    @RedDeprecated(since = "0.5", forRemovalIn = "0.8")
     public static byte clampMaxFirst(final byte input, final byte min, final byte max) {
         return ClampUtil.clampMaxFirst(input, min, max);
     }
@@ -77,7 +78,7 @@ public final class MathUtil {
     /**
      * @deprecated Use {@link ClampUtil#clampTest}
      */
-    @Deprecated
+    @RedDeprecated(since = "0.5", forRemovalIn = "0.8")
     public static byte clampTest(final byte input, final byte min, final byte max) {
         return ClampUtil.clampTest(input, min, max);
     }
@@ -85,7 +86,7 @@ public final class MathUtil {
     /**
      * @deprecated Use {@link ClampUtil#clampMinFirst}
      */
-    @Deprecated
+    @RedDeprecated(since = "0.5", forRemovalIn = "0.8")
     public static short clampMinFirst(final short input, final short min, final short max) {
         return ClampUtil.clampMinFirst(input, min, max);
     }
@@ -93,7 +94,7 @@ public final class MathUtil {
     /**
      * @deprecated Use {@link ClampUtil#clampMaxFirst}
      */
-    @Deprecated
+    @RedDeprecated(since = "0.5", forRemovalIn = "0.8")
     public static short clampMaxFirst(final short input, final short min, final short max) {
         return ClampUtil.clampMaxFirst(input, min, max);
     }
@@ -101,7 +102,7 @@ public final class MathUtil {
     /**
      * @deprecated Use {@link ClampUtil#clampTest}
      */
-    @Deprecated
+    @RedDeprecated(since = "0.5", forRemovalIn = "0.8")
     public static short clampTest(final short input, final short min, final short max) {
         return ClampUtil.clampTest(input, min, max);
     }
@@ -109,7 +110,7 @@ public final class MathUtil {
     /**
      * @deprecated Use {@link ClampUtil#clampMinFirst}
      */
-    @Deprecated
+    @RedDeprecated(since = "0.5", forRemovalIn = "0.8")
     public static int clampMinFirst(final int input, final int min, final int max) {
         return ClampUtil.clampMinFirst(input, min, max);
     }
@@ -117,7 +118,7 @@ public final class MathUtil {
     /**
      * @deprecated Use {@link ClampUtil#clampMaxFirst}
      */
-    @Deprecated
+    @RedDeprecated(since = "0.5", forRemovalIn = "0.8")
     public static int clampMaxFirst(final int input, final int min, final int max) {
         return ClampUtil.clampMaxFirst(input, min, max);
     }
@@ -125,7 +126,7 @@ public final class MathUtil {
     /**
      * @deprecated Use {@link ClampUtil#clampTest}
      */
-    @Deprecated
+    @RedDeprecated(since = "0.5", forRemovalIn = "0.8")
     public static int clampTest(final int input, final int min, final int max) {
         return ClampUtil.clampTest(input, min, max);
     }
@@ -133,7 +134,7 @@ public final class MathUtil {
     /**
      * @deprecated Use {@link dev.redstudio.redcore.math.MathUtil#absolute}
      */
-    @Deprecated
+    @RedDeprecated(since = "0.5", forRemovalIn = "0.8")
     public static float absolute(final float input) {
         return dev.redstudio.redcore.math.MathUtil.absolute(input);
     }
@@ -141,7 +142,7 @@ public final class MathUtil {
     /**
      * @deprecated Use {@link dev.redstudio.redcore.math.MathUtil#absolute}
      */
-    @Deprecated
+    @RedDeprecated(since = "0.5", forRemovalIn = "0.8")
     public static double absolute(final double input) {
         return dev.redstudio.redcore.math.MathUtil.absolute(input);
     }
@@ -149,7 +150,7 @@ public final class MathUtil {
     /**
      * @deprecated Use {@link dev.redstudio.redcore.math.MathUtil#absoluteMax}
      */
-    @Deprecated
+    @RedDeprecated(since = "0.5", forRemovalIn = "0.8")
     public static float absoluteMax(final float input1, final float input2) {
         return dev.redstudio.redcore.math.MathUtil.absoluteMax(input1, input2);
     }
@@ -157,7 +158,7 @@ public final class MathUtil {
     /**
      * @deprecated Use {@link dev.redstudio.redcore.math.MathUtil#absoluteMax}
      */
-    @Deprecated
+    @RedDeprecated(since = "0.5", forRemovalIn = "0.8")
     public static double absoluteMax(final double input1, final double input2) {
         return dev.redstudio.redcore.math.MathUtil.absoluteMax(input1, input2);
     }
@@ -165,7 +166,7 @@ public final class MathUtil {
     /**
      * @deprecated Use {@link dev.redstudio.redcore.math.MathUtil#addOrSubtractBasedOnSign}
      */
-    @Deprecated
+    @RedDeprecated(since = "0.5", forRemovalIn = "0.8")
     public static float addOrSubtractBasedOnSign(final float input, final float add) {
         return dev.redstudio.redcore.math.MathUtil.addOrSubtractBasedOnSign(input, add);
     }
@@ -173,7 +174,7 @@ public final class MathUtil {
     /**
      * @deprecated Use {@link dev.redstudio.redcore.math.MathUtil#addOrSubtractBasedOnSign}
      */
-    @Deprecated
+    @RedDeprecated(since = "0.5", forRemovalIn = "0.8")
     public static double addOrSubtractBasedOnSign(final double input, final double add) {
         return dev.redstudio.redcore.math.MathUtil.addOrSubtractBasedOnSign(input, add);
     }
@@ -181,7 +182,7 @@ public final class MathUtil {
     /**
      * @deprecated Use {@link dev.redstudio.redcore.math.MathUtil#round}
      */
-    @Deprecated
+    @RedDeprecated(since = "0.5", forRemovalIn = "0.8")
     public static float round(final float input, final int decimals) {
         return dev.redstudio.redcore.math.MathUtil.round(input, decimals);
     }
@@ -189,7 +190,7 @@ public final class MathUtil {
     /**
      * @deprecated Use {@link dev.redstudio.redcore.math.MathUtil#round}
      */
-    @Deprecated
+    @RedDeprecated(since = "0.5", forRemovalIn = "0.8")
     public static double round(final double input, final int decimals) {
         return dev.redstudio.redcore.math.MathUtil.round(input, decimals);
     }
@@ -197,7 +198,7 @@ public final class MathUtil {
     /**
      * @deprecated Use {@link dev.redstudio.redcore.math.MathUtil#floorToInt}
      */
-    @Deprecated
+    @RedDeprecated(since = "0.5", forRemovalIn = "0.8")
     public static int floorToInt(final float input) {
         return dev.redstudio.redcore.math.MathUtil.floorToInt(input);
     }
@@ -205,7 +206,7 @@ public final class MathUtil {
     /**
      * @deprecated Use {@link dev.redstudio.redcore.math.MathUtil#floorToInt}
      */
-    @Deprecated
+    @RedDeprecated(since = "0.5", forRemovalIn = "0.8")
     public static int floorToInt(final double input) {
         return dev.redstudio.redcore.math.MathUtil.floorToInt(input);
     }
@@ -213,7 +214,7 @@ public final class MathUtil {
     /**
      * @deprecated Use {@link dev.redstudio.redcore.math.MathUtil#lerp}
      */
-    @Deprecated
+    @RedDeprecated(since = "0.5", forRemovalIn = "0.8")
     public static float lerp(final float startValue, final float partialTicks, final float targetValue) {
         return dev.redstudio.redcore.math.MathUtil.lerp(startValue, partialTicks, targetValue);
     }
@@ -221,7 +222,7 @@ public final class MathUtil {
     /**
      * @deprecated Use {@link dev.redstudio.redcore.math.MathUtil#lerp}
      */
-    @Deprecated
+    @RedDeprecated(since = "0.5", forRemovalIn = "0.8")
     public static double lerp(final double startValue, final double partialTicks, final double targetValue) {
         return dev.redstudio.redcore.math.MathUtil.lerp(startValue, partialTicks, targetValue);
     }
@@ -229,7 +230,7 @@ public final class MathUtil {
     /**
      * @deprecated Use {@link dev.redstudio.redcore.math.MathUtil#boolToInt}
      */
-    @Deprecated
+    @RedDeprecated(since = "0.5", forRemovalIn = "0.8")
     public static int boolToInt(final boolean input) {
         return dev.redstudio.redcore.math.MathUtil.boolToInt(input);
     }