+++ mc/1-8---1-12/src/main/java/io/redstudioragnarok/redcore/utils/NetworkUtil.java
@@ -1,18 +1,19 @@
 package io.redstudioragnarok.redcore.utils;
 
+import dev.redstudio.redcore.annotations.RedDeprecated;
 import net.minecraftforge.fml.common.network.simpleimpl.MessageContext;
 
 /**
  * @author Luna Mira Lage (Desoroxxx)
  * @deprecated See methods for details.
  */
-@Deprecated // Todo: Remove in 0.8
+@RedDeprecated(since = "0.5", forRemovalIn = "0.8")
 public final class NetworkUtil {
 
     /**
      * @deprecated Use {@link dev.redstudio.redcore.utils.NetworkUtil#processMessage}
      */
-    @Deprecated
+    @RedDeprecated(since = "0.5", forRemovalIn = "0.8")
     public static void processMessage(final MessageContext messageContext, final Runnable runnable) {
         dev.redstudio.redcore.utils.NetworkUtil.processMessage(messageContext, runnable);
     }