+++ mc/1-8---1-12/src/main/java/io/redstudioragnarok/redcore/utils/OptiNotFine.java
@@ -1,5 +1,6 @@
 package io.redstudioragnarok.redcore.utils;
 
+import dev.redstudio.redcore.annotations.RedDeprecated;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
 
@@ -7,7 +8,7 @@ import net.minecraftforge.fml.relauncher.SideOnly;
  * @author Luna Mira Lage (Desoroxxx)
  * @deprecated See methods for details.
  */
-@Deprecated // Todo: Remove in 0.8
+@RedDeprecated(since = "0.5", forRemovalIn = "0.8")
 @SideOnly(Side.CLIENT)
 @SuppressWarnings("unused")
 public final class OptiNotFine {
@@ -15,7 +16,7 @@ public final class OptiNotFine {
     /**
      * @deprecated Use {@link dev.redstudio.redcore.utils.OptiNotFine#isOptiFineInstalled()}
      */
-    @Deprecated
+    @RedDeprecated(since = "0.5", forRemovalIn = "0.8")
     public static boolean isOptiFineInstalled() {
         return dev.redstudio.redcore.utils.OptiNotFine.isOptiFineInstalled();
     }
@@ -23,7 +24,7 @@ public final class OptiNotFine {
     /**
      * @deprecated Use {@link dev.redstudio.redcore.utils.OptiNotFine#shadersEnabled}
      */
-    @Deprecated
+    @RedDeprecated(since = "0.5", forRemovalIn = "0.8")
     public static boolean shadersEnabled() {
         return dev.redstudio.redcore.utils.OptiNotFine.shadersEnabled();
     }